Preparaz. a: Split/Grow Far Button; fix vari map

+++ src/selection_map.cpp
@@ -1,4 +1,4 @@
 #include <neural-graphics-primitives/selection_map.h>
 #include <Eigen/Dense>
 
-std::map<std::size_t, Eigen::Vector3f> selection_map::privateMap; 
+std::map<std::uint32_t, Eigen::Vector3f> selection_map::privateMap;
